Tidy up paths and class names
