+++ lostphone/js/main.js
@@ -1,9 +1,9 @@
-import config from './config'
+import Config from '/Config'
 
 // Sure there's another aesthetic way
 // to launch the game
 
 // --- Booting up
 window.addEventListener('load', () => {
-  var game = new Phaser.Game(config);
+  var game = new Phaser.Game(Config);
 });
