+++ lostphone/js/prefabs/sprite.js
@@ -1,4 +1,4 @@
-import { assetsDPR } from '../config.js'
+import { assetsDPR } from '/Config.js'
 
 class Sprite extends Phaser.GameObjects.Sprite {
   constructor (scene, x, y, texture, frame) {
